+++ src/lib/pages/codes/components/CodesTable.tsx
@@ -6,21 +6,23 @@ import {
   Tr,
   Th,
   Td,
-  Button,
   TableContainer,
   Heading,
   HStack,
   VStack,
   Text,
   Box,
+  Flex,
 } from "@chakra-ui/react";
 import { useWallet } from "@cosmos-kit/react";
 import { useRouter } from "next/router";
 import type { ReactNode } from "react";
 import { MdSearchOff } from "react-icons/md";
 
+import { InstantiateButton } from "lib/components/button/InstantiateButton";
 import { ExplorerLink } from "lib/components/ExplorerLink";
 import { RemoveCode } from "lib/components/modal/code/RemoveCode";
+import { PermissionChip } from "lib/components/PermissionChip";
 import { DisconnectedState } from "lib/components/state/DisconnectedState";
 import type { CodeInfo } from "lib/types";
 
@@ -103,12 +105,12 @@ const TableHead = () => {
         }}
       >
         <Th width="10%">Code ID</Th>
-        <Th width="45%">Code Description</Th>
+        <Th width="35%">Code Description</Th>
         <Th width="10%" textAlign="center">
           Contracts
         </Th>
         <Th width="15%">Uploader</Th>
-        <Th width="20%" />
+        <Th width="30%">Permission</Th>
       </Tr>
     </Thead>
   );
@@ -117,8 +119,8 @@ const TableHead = () => {
 const TableRow = ({ code, isRemovable }: CodesRowProps) => {
   const router = useRouter();
 
-  const goToInstantiate = () => {
-    router.push({ pathname: "/instantiate", query: { "code-id": code.id } });
+  const goToCodeDetails = () => {
+    router.push({ pathname: `/code/${code.id}` });
   };
 
   return (
@@ -128,15 +130,30 @@ const TableRow = ({ code, isRemovable }: CodesRowProps) => {
       _hover={{
         bg: "gray.900",
       }}
+      cursor="pointer"
+      onClick={goToCodeDetails}
     >
       <Td width="10%" color="primary.main">
-        <ExplorerLink value={code.id.toString()} canCopyWithHover />
+        <ExplorerLink
+          type="code_id"
+          value={code.id.toString()}
+          canCopyWithHover
+        />
       </Td>
-      <Td width="45%">
+      <Td width="35%">
         <CodeDescriptionCell codeId={code.id} description={code.description} />
       </Td>
       <Td width="10%" textAlign="center">
-        {code.contracts}
+        <Text
+          variant="body2"
+          onClick={(e) => e.stopPropagation()}
+          cursor="text"
+          w="fit-content"
+          m="auto"
+          px={2}
+        >
+          {code.contracts}
+        </Text>
       </Td>
       <Td width="15%">
         <ExplorerLink
@@ -145,15 +162,23 @@ const TableRow = ({ code, isRemovable }: CodesRowProps) => {
           canCopyWithHover
         />
       </Td>
-      <Td width="20%">
-        <HStack>
-          <Button variant="outline-gray" size="sm" onClick={goToInstantiate}>
-            Instantiate
-          </Button>
-          {isRemovable && (
-            <RemoveCode codeId={code.id} description={code.description} />
-          )}
-        </HStack>
+      <Td width="30%">
+        <Flex justify="space-between" align="center">
+          <PermissionChip
+            instantiatePermission={code.instantiatePermission}
+            permissionAddresses={code.permissionAddresses}
+          />
+          <HStack onClick={(e) => e.stopPropagation()}>
+            <InstantiateButton
+              instantiatePermission={code.instantiatePermission}
+              permissionAddresses={code.permissionAddresses}
+              codeId={code.id}
+            />
+            {isRemovable && (
+              <RemoveCode codeId={code.id} description={code.description} />
+            )}
+          </HStack>
+        </Flex>
       </Td>
     </Tr>
   );